Refactor: Setsubmitaction from editProfile. Remove: Utils.js and export config into index.js. Deleted unused functions.
