+++ src/utils/constants.js
@@ -1,0 +1,10 @@
+const config = {
+  formSelector: ".modal__form",
+  inputSelector: ".modal__edit",
+  submitButtonSelector: ".modal__button",
+  inactiveButtonClass: "modal__button_disabled",
+  inputErrorClass: ".modal__error",
+  errorClass: ".modal__error_visible",
+};
+
+export { config };
